--- src/main/java/ListCommand.java
@@ -1,8 +1,0 @@
-public class ListCommand extends Command{
-
-    @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage) throws MonaException {
-        ui.showAllTasks(tasks.getTasks());
-    }
-
-}
