+++ src/main/java/command/AddTaskCommand.java
@@ -1,3 +1,10 @@
+package command;
+
+import storage.Storage;
+import task.Task;
+import task.TaskList;
+import ui.Ui;
+
 public class AddTaskCommand extends Command {
     private Task task;
