+++ src/main/java/command/ByeCommand.java
@@ -1,3 +1,9 @@
+package command;
+
+import storage.Storage;
+import task.TaskList;
+import ui.Ui;
+
 public class ByeCommand extends Command {
 
     @Override
