+++ src/main/java/command/DeleteCommand.java
@@ -1,3 +1,11 @@
+package command;
+
+import exception.MonaException;
+import storage.Storage;
+import task.Task;
+import task.TaskList;
+import ui.Ui;
+
 public class DeleteCommand extends Command {
 
     private int index;
