+++ src/main/java/task/Deadline.java
@@ -1,25 +1,29 @@
+package task;
+
 import java.time.LocalDateTime;
+
+import exception.MonaException;
 
 public class Deadline extends TimedTask {
     protected LocalDateTime doneBy;
 
     public Deadline(String description, String doneBy) throws MonaException {
         super(description);
-        this.doneBy = parseDateTime(doneBy);
+        this.doneBy = TimedTask.parseDateTime(doneBy);
     }
 
     public Deadline(String description, boolean isDone, String doneBy) throws MonaException {
         super(description, isDone);
-        this.doneBy = parseDateTime(doneBy);
+        this.doneBy = TimedTask.parseDateTime(doneBy);
     }
 
     @Override
     public String toString() {
-        return "[D]" + super.toString() + " (by: %s)".formatted(formatDateTime(doneBy));
+        return "[D]" + super.toString() + " (by: %s)".formatted(TimedTask.formatDateTime(doneBy));
     }
 
     @Override
     public String toSaveFormat() {
-        return "D | %s | %s | %s".formatted(isDone ? "1" : "0", description, doneBy.format(INPUT_FORMATTER));
+        return "D | %s | %s | %s".formatted(isDone ? "1" : "0", description, doneBy.format(TimedTask.INPUT_FORMATTER));
     }
 }
